+++ packages/e2e/e2e-docs/tests/page-crawler.spec.ts
@@ -2,14 +2,7 @@
 /* eslint-disable playwright/expect-expect */
 import { test } from '@kadena-dev/e2e-base/src/fixtures/shared/test.fixture';
 
-const pages = [
-  'learn',
-  'build',
-  'deploy',
-  'participate',
-  'tutorials',
-  'reference',
-];
+const pages = ['learn', 'build', 'deploy', 'participate', 'reference'];
 for (const pageToCheck of pages) {
   test(`Validate menu structure for: ${pageToCheck}`, async ({
     page,