+++ packages/apps/dev-wallet/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @kadena/dev-wallet
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies \[de4fb13]
+  - @kadena/kode-icons\@0.2.0
+  - @kadena/kode-ui\@0.14.3
+
 ## 0.3.3
 
 ### Patch Changes